Extend `gem` DSL with a `force_ruby_platform` option
It's useful, specially when we enable the `specific_platform` setting by
default to be able to opt-out of the standard platform priority and
always force the ruby variant for specific gems.

This is the "per gem" equivalent of the `force_ruby_platform` setting.

This can be handy (assuming the pure ruby variant works just fine) when:

* You're having issues with the platform specific variant.

* The platform specific variant does not yet support a newer ruby (and thus has
  a `required_ruby_version` upper bound), but you still want your Gemfile{.lock}
  files to resolve under that ruby.